requirements.txt
torch
torchvision
pandas
numpy
tqdm
Pillow
insightface
onnxruntime
transformers